Fix plain omp tool
