+++ critical-core.cpp
@@ -14,6 +14,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "criticalPath.h"
+#include "errorhandler.h"
 #include "parse_flags.h"
 #include <time.h>
 #ifdef USE_MPI
@@ -24,6 +25,63 @@ extern "C" void __cxa_pure_virtual() {
   assert(false && "Pure virtual function must not be called");
   abort();
 }
+
+#ifdef USE_ERRHANDLER
+#include <execinfo.h>
+#include <signal.h>
+#include <sys/types.h>
+#include <unistd.h>
+
+#define CALLSTACK_SIZE 20
+static void print_stack(void) {
+  int nptrs;
+  void *buf[CALLSTACK_SIZE + 1];
+  nptrs = backtrace(buf, CALLSTACK_SIZE);
+  backtrace_symbols_fd(buf, nptrs, STDOUT_FILENO);
+}
+
+static void set_signalhandlers(sighandler_t handler) {
+  signal(SIGSEGV, handler);
+  signal(SIGINT, handler);
+  signal(SIGHUP, handler);
+  signal(SIGABRT, handler);
+  signal(SIGTERM, handler);
+  signal(SIGUSR2, handler);
+  signal(SIGQUIT, handler);
+  signal(SIGALRM, handler);
+}
+
+void disable_signalhandlers() { set_signalhandlers(SIG_DFL); }
+
+void mySignalHandler(int signum) {
+  disable_signalhandlers();
+  printf("pid %i caught signal nr %i\n", getpid(), signum);
+  if (signum == SIGINT || signum == SIGKILL) {
+    print_stack();
+    _exit(signum + 128);
+  }
+  if (signum == SIGTERM || signum == SIGUSR2) {
+    print_stack();
+    fflush(stdout);
+    sleep(1);
+    _exit(signum + 128);
+  }
+  print_stack();
+
+  printf("Waiting up to %i seconds to attach with a debugger.\n", 20);
+  sleep(20);
+  _exit(1);
+}
+
+void init_signalhandlers() {
+  if (!analysis_flags->stacktrace)
+    return;
+  if (analysis_flags->verbose)
+    fprintf(analysis_flags->output, "Setting signal handlers\n");
+  set_signalhandlers(mySignalHandler);
+}
+
+#endif
 
 double localTimeOffset{0};
 double getTime() {
